CDAP-14022 add bootstrap framework
Added the basic framework for CDAP bootstrap.
A new 'bootstrap.file' config setting was added, which is expected
to point to a file that contains a list of bootstrap steps to
perform, such as loading system artifacts and creating the default
namespace. Each type of step expects different arguments in a
different format.

Added types for creating the native profile, creating the default
namespace, creating a system profile, and loading system artifacts.
More types will be added in the future.